Updating models based on PR comment
