--- EyeSpeak/Features/Presets/Models/Keyboard.swift
@@ -1,14 +1,0 @@
-//
-//  Keyboard.swift
-//  EyeSpeak
-//
-//  Created by Jesse Morgan on 3/3/20.
-//  Copyright © 2020 WillowTree. All rights reserved.
-//
-
-import Foundation
-
-struct KeyboardKeys {
-    static var qwerty = "QWERTYUIOPASDFGHJKL'ZXCVBNM,.?"
-    static var alphabetical = "ABCDEFGHIJKLMNOPQRSTUVWXYZ',.?"
-}
